+++ examples/into_response.rs
@@ -1,4 +1,7 @@
-use foxhole::{run, sys, systems::Endpoint, Get, IntoResponse, Response, Route};
+use foxhole::{
+    resolve::{Endpoint, Get},
+    run, sys, IntoResponse, Response, Route,
+};
 
 // This is a reimplementation of the provided `Html` type.
 struct Html(String);