+++ src/tasks.rs
@@ -19,7 +19,7 @@ use crate::{
     routing::Route,
     sequential_writer::{self, SequentialWriter},
     type_cache::TypeCacheShared,
-    MaybeIntoResponse,
+    Action,
 };
 
 const MIN_THREADS: usize = 4;
@@ -141,9 +141,7 @@ impl Task for RequestTask {
             }
         }
 
-        let _ = self
-            .writer
-            .send(&404u16.maybe_response().unwrap().into_raw_bytes());
+        let _ = self.writer.send(&404u16.action().unwrap().into_raw_bytes());
     }
 }
 
@@ -253,11 +251,10 @@ impl TaskPool {
     {
         self.shared.pool.lock().unwrap().push_back(Box::new(task));
 
-        
         if self.shared.waiting_tasks.load(Ordering::Acquire) < MIN_THREADS {
             self.spawn_thread(true);
         }
-        
+
         // FIXME potential race condition where thread is not yet waiting on the condvar. Support
         // spawning threads with a task directly in the case there are 0 waiting
         self.shared.condvar.notify_one();